Add test and migrate spm test to pytest
